microservices/requirements.txt
redis==4.5.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.7
